(scss): set up scss structure
A folder named style which contains partials for mixins and variables as well as the main style.scss file was created in the /client folder. Some mixins and variables were defined. Body style was defined in style.scss. /imports/ui/home/_home.scss was created and a style was added to it.