(packages/version) update, minor changes
in playing around phase, trying to get meteor to do what i want it to. havent worked on this in almost a week, i think i did a reset head~1 and ended up with some uncommitted stuff? git status shows changes to package and version folders. meteor --version showed 1.3.1. starting this branch to mess around on, committing those changes because they should be harmless, and to check them out on github.