(functions): get platforms working decently
been working on getting this system to conserve space working... the FilledSpaceModel object keeps track of the space that's occupied. the fsmodel has arrays of 'platforms' which can be thought of like skylines, and the platforms are the rooftops. they let the program figure out where the next text box can fit in.

i set up the updateTopLine and it is working ok. did a ton of coding here, too much to go over, but its all based on getting this method of fitting textboxes together efficiently to work