(functions) replace platform based FSM with box-based FSM
in order to allow a complex outline for the filled space model, so that gnarly branches of the discussion can grow and get end heavy, i'm ditching the 'platform'-based FilledSpaceModel i was using. constructor for line was made, including side property to mark the outside. fsm constructor hollowed out pretty much, it has an addMotion method, a reset() method, and it initializes the lineArray. I redid the drawFSM function, it draws the fsm and marks the outside of each line. i also redid some of the measurement.js functions.