(fsm): implement better way of removing orphan lines
when there are two crossing lines, after it finds out what side the outside is, it sets a variable marking the side of the margin box that should be removed, and returns it, then trimOverlap finds it in lineArray and removes it. also renamed closestCrossingLineOnInsideSide to getOtherLine, and insideCrossingLine to otherCrossingLine. also removed some extraneous conditionals from getOtherLine, because there should only ever be a maximum of 2 crossing lines.