(grow): get checkIfOverEdge() working
removed some tech debt in reactive.js, elicitorIsMotion and Settings were unneeded as parameters. renamed some things, checkIfOverEdge now keeps track of how much further beyond the canvas's initial bounds the debate tree extends, it is an array of 4 numbers. the function that checks is now called checkIfOverEdge. code was added to do the comparisons in that function.