(posts): method updates date discussed based on motion field now
a response to a response triggers the original motion of the discussion having its dateDiscussed field updated by getting the id of the motion from the motion field of the post. also updated add-ten in dev-tool so it works, and removed some unneeded things from dev-tool. also updated the dummy data method so it provides motion fields